pyproject.toml
[project]
name = "personalfinanceanalyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "Bradley Brenning",email = "Brenbra132@gmail.com"}
]

[tool.poetry]
# Use a minimal Poetry configuration to manage dependencies. Keep package
# metadata simple and put dependencies under the canonical Poetry keys so
# `poetry install --with dev` works in CI. Avoid specifying `dependencies`
# as a TOML array named 'dependencies' which was causing poetry to parse a
# legacy format and fail with "'list' object has no attribute 'keys'".
name = "personalfinanceanalyzer"
version = "0.1.0"
description = "Personal Finance Analyzer - minimal Streamlit app"
authors = ["Bradley Brenning <Brenbra132@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<4"
streamlit = "^1.56.0"
sqlalchemy = "^2.0.49"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.2"
pandas = "^3.0.2"
reportlab = "^4.2.5"
sentry-sdk = "^2.58.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
ruff = "^0.6.9"
pre-commit = "^4.5.1"
mypy = "^1.20.1"
pytest-playwright = "^0.3.0"
playwright = "^1.35.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = false
warn_unused_configs = true
exclude = ["tests"]
